requirements.txt
# Core dependencies
opencv-python==4.9.0.80
numpy==1.26.4
google-cloud-aiplatform==1.45.0
google-generativeai==0.5.2
transformers==4.40.1
deepface==0.0.89
python-dotenv==1.0.1
aiohttp==3.9.5

# Required for audio processing
ffmpeg-python==0.2.0

# Note: google-adk must be installed separately from Google's sources
# Haarcascades are included with opencv-python
# Google Cloud credentials must be configured via gcloud
